requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
tiktoken
youtube-transcript-api